Add files via upload
I attempted to perform NGS data analysis for variant calling on a randomly selected sample of Drosophila by utilizing publicly available datasets. The primary objective of this project was to gain hands-on experience with NGS data analysis techniques and tools for variant calling. In order to accomplish this goal, I followed a basic work pipeline that involved performing quality control, alignment, and variant calling using relevant software tools. By executing this pipeline, I was able to identify and annotate the genetic variations present in the sample, which is a crucial step towards understanding the underlying biology of the species under investigation. This project served as an excellent opportunity for me to explore and refine my skills in the field of NGS data analysis, and I have documented the entire process in a publicly accessible GitHub repository for others to learn from and replicate.